fix(search-streamer): Added document truncation
Extremely large results (larger than 100 MB) gave 413 erros in elastic
Increasing the maximum request size in elastic fixed the issue, but
truncation was also added in case any results are just too large for
elastic to accept. The size to truncate to per document is about 100MB
(as defined in the Thorium config/CRD). Each result is proportionally
truncated based on the number of bytes it takes up in the document to
avoid truncating smaller results to 0.