Added Tagged Container Releases
This adds container builds for tagged branches that use a semantic version tag
in the form of X.Y.Z*. Github pages docs will only build and deploy for the
mainline branch. This also fixes a broken reference to Thorctl architecture
docs. This also fixes windows binary builds for thorctl, thorium-agent, and
the thorium-reactor.

Closes #5