fix(scaler)!: Fixed issue where the scaler would only do fair share scheduling
If fairs hare was enabled then it tended to consume all spawn slots and
starve out the deadline pool. This would lead to poor cluster utilization
in cases when jobs complete very quickly.

This removes the max_sway part of the Thorium CRD in favor of a
SpawnSlots based structure.