fix(cart-rs): Fixed incorrect header format
The 'bincode' crate's 'encode_into_slice' function does not
automatically advance the buffer it writes to because it creates
a new slice to write to when encoding. That meant that after
encoding the reserved size, the buffer's location was still
right after the version number, so the encryption key was
written to the wrong place. cart-rs now manually tracks where
to write the key to avoid this issue.