fix(api): Fixed parsing error when tag key begins with number
The error occurred because serde-qs got confused when seeing a
percent-encoded open bracket from the tags query parameter, then seeing
a tag key with a number right after the encoding. This caused a
ParseIntError. Upgrading serde-qs to latest release candidate fixes this
issue. Activated the "default_to_form_encoding" feature in serde-qs to
support percent-encoded brackets by default (the Web UI percent-encodes
brackets so we need to support it).