feat(api): Added support for Reaction cache
This allows reactions to have a cache of data and files shared across
images. The Thorium agent will automatically sync this cache at the end
of every job for future jobs to use. This also fixes an issue with some
MCP tools calls returning invalid structured content. This also fixes an
issue with burstable being required in the thorium config. This also
fixes an issue with generators never waking up after sleeping.