fix(api)!: Fixed issue where trees would be missing branches
This was due to filtering out some nodes from being crawled incorrectly.
This also resolves an issue with the scaler not scheduling workers due
to an issue with determining if we have enough burstable resources
available. This also adds support for deserializing cpu and storage
resources from integers, strings, and floats (only cpu supports floats).